+++ src/core/mod.rs
@@ -1,3 +1,4 @@
+/// Spice API for accessing kernel data.
 mod kernel;
 
 pub use self::kernel::*;