+++ src/lib.rs
@@ -3,7 +3,7 @@
 
 WOW! The complete NASA/NAIF Spice toolkit is actually usable on Rust.
 
-## Using **kalast**
+## Using **rust-spice**
 
 Simply add the following to your `Cargo.toml` file:
 
@@ -12,23 +12,26 @@ Simply add the following to your `Cargo.toml` file:
 rust-spice = "0.3.5"
 ```
 
-Useful functionalities of **rust-spice** are grouped in the root module `spice::`.
+Rust layered Spice functions of **rust-spice** are grouped in the root module `spice::` and all wrapper functions in `spice::c::`.
 
 ## **rust-spice** basics
 
 ```.ignore
 use spice;
 
-// To load and unload a metakernel with the Rust layer.
-spice::load("/path/to/metakernel.mk");
-spice::unload("/path/to/metakernel.mk");
-
-// To load and unload a metakernel with the C wrapper.
+// Example using binded C Spice.
 unsafe {
-    let kernel = CString::new("/path/to/metakernel.mk").unwrap();
-    spice::c::furnsh_c(kernel.as_ptr() as *mut _);
-    spice::c::unload_c(kernel.as_ptr() as *mut _);
+    let kernel = CString::new("/path/to/metakernel.mk").unwrap().into_raw();
+    spice::c::furnsh_c(kernel);
+    spice::c::unload_c(kernel);
 }
+
+// Using the nice Rust interface.
+let mut kernel = spice::Kernel::new("/path/to/metakernels.mk")?;
+kernel.unload()?;
+
+// If you are curious reading the sources, do not be tempted to use directly `spice::load(..)`.
+// The Kernel type adds extra error checks.
 ```
 
 ## In action
@@ -40,28 +43,27 @@ use itertools::multizip;
 use spice;
 
 // Define the system.
-// The metakernel is automatically loaded.
-let mut system = spice::System::new(
-    "rsc/data/hera_PO_EMA_2024.tm", // kernel
-    "J2000",                        // frame
-    "HERA",                         // observer
-    "DIMORPHOS",                    // target
-    "2027-MAR-23 16:00:00",         // start date
-    129.0 * spice::DAY(),           // duration
-    "NONE",                         // aberration correction
-);
+let mut system = spice::SystemBuilder::default()
+    .kernel("/path/to/metakernels.mk")?
+    .frame("J2000")
+    .observer("HERA")
+    .target("DIMORPHOS")
+    .start_date("2027-MAR-23 16:00:00")
+    .duration(129.0 * spice::DAY)
+    .aberration_correction("NONE")
+    .build()?;
 
 // Define the time step.
-let time_step = 1.0 * spice::DAY();
+let time_step = 1.0 * spice::DAY;
 
 // Get all the times string-formatted.
 let times = system.times_formatted(time_step);
 
 // Get all the positions at the times.
-let mut positions = system.positions(time_step);
+let positions = system.positions(time_step);
 
 // Get the distances from the positions.
-let distances = spice::compute_distances(&mut positions);
+let distances = spice::distance(positions);
 
 // Display
 for (time, distance) in multizip((times.iter(), distances.iter())) {
@@ -69,11 +71,10 @@ for (time, distance) in multizip((times.iter(), distances.iter())) {
 }
 
 // Always unload the kernels.
-system.unload();
-
+system.unload()?;
 ```
 
-You can also read other [examples](https://github.com/GregoireHENRY/rust-spice/examples).
+You can also read other [examples](https://github.com/GregoireHENRY/rust-spice/tree/main/examples).
 
 ## Rust layer in development
 
@@ -82,9 +83,9 @@ You can also read other [examples](https://github.com/GregoireHENRY/rust-spice/e
 The complete NASA/NAIF C Spice toolkit is accessible under the module `spice::c`. It has been
 generated using [`bindgen`](https://docs.rs/bindgen/0.57.0/bindgen/).
 
-If you feel some troubles to use the C interface, you can visit the sources of [`System`]. Most
+If you feel some troubles to use the C interface, you can visit the sources of [`core`]. Most
 conversions from pointers to list of doubles or chars, back and forth, are done there. A complete
-guide will be written if needed.
+guide can be written if needed.
 
 ### The damn Rust layer
 
@@ -93,6 +94,10 @@ is not automated, I have to do it by hand. I started from my needs, basically ge
 start to end date of target in referential. I will implemented a nice Rust layer more and more when
 I will be using other functions. You can submit issues if you want a layer to some functionalities,
 and we will implement it immediately.
+
+The Rust layer for the most Used C Spice API for accessing kernel data is located in [`core`].
+
+The type [`System`] provides some tools, built on top of spice.
 */
 
 #![allow(non_upper_case_globals)]
@@ -104,18 +109,17 @@ extern crate log;
 extern crate nalgebra as na;
 extern crate simplelog;
 
-/// Unsafe C Spice wrapped.
+/// Complete NASA/NAIF C Spice binded functions, very unsafe.
 pub mod c {
     include!(concat!(env!("OUT_DIR"), "/bindings.rs"));
 }
-
-/// Core features.
+/// The Rust layer to ease the use of the wrapper.
 pub mod core;
-/// Generic functions for math, physics, matrix operations.
+/// Tools developped on top of Spice for even easier usage of the library.
+pub mod spicetools;
+/// Generic functions for geometry and matrix operations.
 #[macro_use]
 pub mod toolbox;
-/// Some tools developped for quicker Spice usage.
-pub mod spicetools;
 
 pub use crate::core::*;
 pub use crate::spicetools::*;