+++ wrapper.h
@@ -1,1 +1,0 @@
-#include "spicetools/includes/spicetools.h"